requirements.txt
# =============================================================================
# ChatTLA — Dependency Manifest
# =============================================================================
# Fine-tuning gpt-oss-20b for TLA+ specification generation.
# Install with:  pip install -r requirements.txt
# GPU:           Quadro RTX 8000, CUDA 12.8, CUDA_VISIBLE_DEVICES=1
# =============================================================================

# --- Core ML / Fine-tuning --------------------------------------------------
# transformers>=4.55 is required for Mxfp4Config and gpt-oss architecture support
torch>=2.6.0
transformers>=4.55.0
peft>=0.17.0          # LoRA / QLoRA adapter support
trl>=0.20.0           # SFTTrainer for supervised fine-tuning
accelerate>=1.3.0     # device_map, gradient checkpointing
bitsandbytes>=0.45.0  # INT8/INT4 fallback quantization (non-MXFP4 paths)
datasets>=3.3.0       # HuggingFace dataset utilities
safetensors>=0.5.0    # weight serialisation

# --- Harmony Format (mandatory for gpt-oss) ---------------------------------
openai-harmony==0.0.8  # Applies the <|start|>/<|end|>/<|message|>/<|return|> token format

# --- Inference --------------------------------------------------------------
ollama>=0.4.0          # Ollama Python client — talks to local ollama daemon
openai>=1.63.0         # OpenAI-compatible client (Ollama exposes /v1)

# --- Data / Scraping --------------------------------------------------------
PyGithub>=2.5.0        # GitHub API v3 for Tier-2 spec scraping
requests>=2.32.0
beautifulsoup4>=4.13.0
pypdf>=5.1.0           # Extract TLA+ snippets from academic PDFs
datasketch>=1.6.5      # MinHash LSH for near-deduplication (Jaccard > 0.8)

# --- Experiment Tracking / Orchestration ------------------------------------
mlflow>=2.20.0
zenml>=0.72.0
jupyter>=1.1.0
ipywidgets>=8.1.0
matplotlib>=3.10.0
pandas>=2.2.0
tqdm>=4.67.0

# --- Utilities --------------------------------------------------------------
python-dotenv>=1.0.0   # Load .env for GITHUB_TOKEN_1..4, OLLAMA_HOST, etc.
pyyaml>=6.0.2
rich>=13.9.0            # Pretty CLI output for pipeline runs
typer>=0.15.0           # CLI entry points for scraper and training scripts
